xcc Publicrelease 4
Rather long nighttime coding session lead to a strange decision - upload
two packages instead of one.
Publicrelease 4 is the stable branch (as stable as it can be called),
while the Interim 1 is the unstable branch with experimental support for
typedefs and variable declarations with commas in them -
declaring multiple types/variables at once. Interim 1 is also
known broken, maybe someone is willing to fix it, idk.