xcc Publicrelease 5 Interim 1
Currently working on some trickery related to type expressions (look at
internals to see what it is, handle to it is tLexem_Typeexpression) and
met an issue at LxParseType that has just got another bypass, this time
for a just type and nothing else.
Temporairly fixed by killing 'Typeexpressionbypass' in LxParseExpression
and the Nullexpression bypass in LxParseType.

Writing a lexer without going for Lexer hack is going to be complicated
- parse some things one way (from LxParseExpression), other things other
way (from LxParseTypeexpression), and some other things are just invalid
and you need to decide what, when and how.

And I'm *still* confusing LxParseExpression and LxParseTypeexpression.