Change the way standard library is generated
This commit changes `stdlib` rule in Makefile.posttemplate to be a shell
script snippet that iterates over archdefs for which a directory is created
in `./configs` and tries to compile libc for them.

There is now no need to do `make install` before `make stdlib` - configdir
for assembler is specified to be the `./configs` directory in repo and the
xcc-assembler itself is called as `./xcc-assembler`, running from directory
it was just built to.

So now you can build xcc as `make build stdlib install` and it'll do
everything needed.